requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
channels==4.1.0
daphne==4.1.2
django-cors-headers==4.4.0
pillow==11.1.0

# Dependency inti Django (akan terpasang otomatis, tapi boleh dipin):
asgiref==3.8.1
sqlparse==0.5.3
typing_extensions==4.12.2
